comments within widget sagas
