+++ src/packages/core/src/constants.ts
@@ -44,7 +44,7 @@ export const sagaEvents = {
   DATA_FLOW_UPDATE_WIDGET: `${APP_NAMESPACE}/SAGAS/DATA_FLOW/updateWidget`,
   DATA_FLOW_CONFIGURATION_UPDATE: `${APP_NAMESPACE}/SAGAS/DATA_FLOW/configurationUpdate`,
   DATA_FLOW_UPDATE_HOOK_STATE: `${APP_NAMESPACE}/SAGAS/DATA_FLOW/updateHookState`,
-  DATA_FLOW_UNMOUNT: `${APP_NAMESPACE}/SAGAS/DATA_FLOW/UNMOUNT`
+  DATA_FLOW_RESTORED: `${APP_NAMESPACE}/SAGAS/DATA_FLOW/RESTORED`
 };
 
 export const reduxActions = {
